Added basic validation of params, and began exception handling
Did not yet add validation for the file IO helper methods in the process api. I felt this would be very closely tied to handling exceptions and would require the most in-depth validation, so I just wanted to wait for a bit. I also added validation within the custom request classes that get passed as params to public api methods.