ensured all exceptions are handled gracefully
Added really nice error handling to ensure all exceptions that could happen, both expected and unexpected, are handled properly and described to the user in some cases. Also added a bit of validation to the private file helper methods in the process api. Specifically, just checking if the files exist before trying to read or write. Also added the checkpoint5testsuite to the project.