Addressed checkpoint 7 feedback
Fixed an issue with my ProcessAPI that led to errors when tryign to run the NetworkClient using in memory inputs. (Not sure how i missed this issue initally, very poor testing on my part). But it turned out I was returning a null message in hte LoadResponse for CUSTOM resources in my ProcessAPI which was cuasing a null pointer when we tried to access that message field. I was returning a non-null response for loading File inputs so thats why it was only happening for in memory inputs. I just changed the regular process api to return an actual message for sucessfully loading custom resources and that fixed the isssue.