Capstone (#35)
* BigInteger

Made everything work with any-sized integer that's larger than 0.

* Finished Capstone

Added a really cool login functionality using a SQLite database.

Got in a huge fight with my build.gradle, trying to add the SQLite jar as a dependancy, ruined everything, so I had to change a large portion of it.

After doing a lot of testing, everything looks good. All tests are passing. The gRPC services support the login and logout functionality and so forth.

* checkstyle fixes

* checkstyle